+++ src/client/clientnetwork.hpp
@@ -21,7 +21,7 @@ public:
     std::deque<ServerEvent>& getQueuEvent();
     void login(std::string pseudo,std::string password);
     void askRegister(std::string pseudo, std::string password);
-    void createProject(std::string role, sf::Vector2u size, float scale);
+    void createProject(std::string role, sf::Vector2u size, uint scale);
     void getProjectData(int project_id);
     void getUsersProjects(long long id);
     void getProjectList();
